requirements.txt
Django==2.2
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
